requirements/requirements.txt
# Mastermind Framework Base Dependencies
pyyaml
python-dotenv
requests
flask
rich
